Feature/tgyuu/#76

+++ feature/survey/src/main/java/com/wap/wapp/feature/survey/navigation/SurveyNavigation.kt
@@ -45,7 +45,7 @@ fun NavGraphBuilder.surveyNavGraph(
             viewModel = hiltViewModel(),
             onSubmitButtonClicked = navigateToSurvey,
             onBackButtonClicked = navigateToSurvey,
-            eventId = eventId.toInt(),
+            eventId = eventId,
         )
     }
 }
